requirements.txt
streamlit>=1.36.0
requests>=2.32.0
Pillow>=10.4.0
